+++ src/renderer/modules/auth-service/store/auth-service-store.d.ts
@@ -1,8 +1,0 @@
-export type IAuthServiceStore = {
-  sid?: string;
-  jwsHbaSignature?: string;
-  jwsSmcbSignature?: string;
-  challengePath: string;
-  challenge: string;
-  authRequestPath?: string;
-};